+++ src/vector_server.cpp
@@ -157,7 +157,8 @@ void Vector_Server::handle_client(int client_fd)
             if ((command.rfind("INSERT", 0)) == 0) // INSERT ID DIMS key1=abc key2=def key3=xyz F1 F2 F3 ... Fn
             {
                 Vector v;
-                Parse_result results = insert_parsing(v, command);
+                DB_entry en;
+                Parse_result results = parser.insert_parsing(v, command);
                 if (!results.success)
                 {
                     send(client_fd, results.message.data(), results.message.length(), 0);
@@ -194,7 +195,7 @@ void Vector_Server::handle_client(int client_fd)
             {
                 Vector query_v;
                 size_t top_k = 0;
-                Parse_result results = query_parsing(query_v, top_k, command);
+                Parse_result results = parser.query_parsing(query_v, top_k, command);
                 if (!results.success)
                 {
                     send(client_fd, results.message.data(), results.message.length(), 0);
@@ -253,7 +254,7 @@ void Vector_Server::handle_client(int client_fd)
             {
                 std::string id = "";
                 Parse_result results;
-                results = delete_parsing(id, command);
+                results = parser.delete_parsing(id, command);
                 if (!results.success)
                 {
                     send(client_fd, results.message.data(), results.message.length(), 0);
@@ -291,7 +292,7 @@ void Vector_Server::handle_client(int client_fd)
             else if ((command.rfind("SAVE", 0)) == 0) // SAVE
             {
                 Parse_result results;
-                results = save_parsing(command, 0);
+                results = parser.save_parsing(command, 0);
                 if (!results.success)
                 {
                     send(client_fd, results.message.data(), results.message.length(), 0);
@@ -306,7 +307,7 @@ void Vector_Server::handle_client(int client_fd)
             else if ((command.rfind("LOAD", 0)) == 0) // LOAD
             {
                 Parse_result results;
-                results = save_parsing(command, 1);
+                results = parser.save_parsing(command, 1);
                 if (!results.success) // save and load -> 4 chars same logic
                 {
                     send(client_fd, results.message.data(), results.message.length(), 0);